Move FieldConfigIndex to commons module
Search Service should also use FieldConfigIndex when parsing parameters
