+++ .gitignore
@@ -29,6 +29,10 @@ hs_err_pid*
 **/.classpath
 **/.factorypath
 
+# idea files
+.idea
+*.iml
+
 # maven
 target
 bin
