+++ indexer-service/src/main/java/de/cxp/ocs/indexer/IndexItemConverter.java
@@ -4,6 +4,7 @@ import java.util.Map.Entry;
 import java.util.function.Function;
 
 import de.cxp.ocs.config.Field;
+import de.cxp.ocs.config.FieldConfigIndex;
 import de.cxp.ocs.config.FieldConfiguration;
 import de.cxp.ocs.indexer.model.DataItem;
 import de.cxp.ocs.indexer.model.IndexableItem;
