+++ pom.xml
@@ -6,7 +6,7 @@
 	<groupId>de.cxp.ocs</groupId>
 	<artifactId>ocs-parent</artifactId>
 	<packaging>pom</packaging>
-	<version>0.5</version>
+	<version>0.6</version>
 	<name>SearchHub Services - Parent</name>
 
 	<modules>
@@ -27,7 +27,7 @@
 		<dockerfile-maven-plugin.version>1.4.13</dockerfile-maven-plugin.version>
 
 		<ocs-api.version>0.5</ocs-api.version>
-		<ocs-commons.version>0.1.1</ocs-commons.version>
+		<ocs-commons.version>0.2</ocs-commons.version>
 
 		<elasticsearch.version>7.5.0</elasticsearch.version>
 		<spring-boot.version>1.5.22.RELEASE</spring-boot.version>
