+++ search-service/pom.xml
@@ -6,7 +6,7 @@
 	<parent>
 		<groupId>de.cxp.ocs</groupId>
 		<artifactId>ocs-parent</artifactId>
-		<version>0.5</version>
+		<version>0.6</version>
 		<relativePath>../</relativePath>
 	</parent>
